requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
yfinance>=0.2.0
scikit-learn>=1.3.0
requests>=2.31.0
pyyaml>=6.0
ta>=0.10.2
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0